refactor(pricing): Co-locate cost computation with pricing data
- Move calculate_cost_from_agent_usage (and its private helpers) from
  utils.py to pricing.py so all pricing-related code lives in one module
- Drop the unused refresh-prices CLI command and the now-orphaned
  force_refresh helper - the 30-min disk TTL handles staleness, and
  rm of the cache file is enough for forced refresh
